feat: big refactor, changing names of underlying class representations of database entries and splitting core, api, and database into separate files
